Remove StorageUtil::BlockLayoutFromSchema to enforce isolation of data layer
