+++ src/storage/sql_table.cpp
@@ -2,8 +2,56 @@
 #include <set>
 #include <vector>
 #include "common/macros.h"
+#include "storage/storage_util.h"
 
 namespace terrier::storage {
+
+SqlTable::SqlTable(BlockStore *const store, const catalog::Schema &schema, const catalog::table_oid_t oid)
+    : block_store_(store), oid_(oid) {
+  // Begin with the NUM_RESERVED_COLUMNS in the attr_sizes
+  std::vector<uint8_t> attr_sizes;
+  attr_sizes.reserve(NUM_RESERVED_COLUMNS + schema.GetColumns().size());
+
+  for (uint8_t i = 0; i < NUM_RESERVED_COLUMNS; i++) {
+    attr_sizes.emplace_back(8);
+  }
+
+  TERRIER_ASSERT(attr_sizes.size() == NUM_RESERVED_COLUMNS,
+                 "attr_sizes should be initialized with NUM_RESERVED_COLUMNS elements.");
+
+  for (const auto &column : schema.GetColumns()) {
+    attr_sizes.push_back(column.GetAttrSize());
+  }
+
+  auto offsets = storage::StorageUtil::ComputeBaseAttributeOffsets(attr_sizes, NUM_RESERVED_COLUMNS);
+
+  ColumnMap col_oid_to_id;
+  // Build the map from Schema columns to underlying columns
+  for (const auto &column : schema.GetColumns()) {
+    switch (column.GetAttrSize()) {
+      case VARLEN_COLUMN:
+        col_oid_to_id[column.GetOid()] = col_id_t(offsets[0]++);
+        break;
+      case 8:
+        col_oid_to_id[column.GetOid()] = col_id_t(offsets[1]++);
+        break;
+      case 4:
+        col_oid_to_id[column.GetOid()] = col_id_t(offsets[2]++);
+        break;
+      case 2:
+        col_oid_to_id[column.GetOid()] = col_id_t(offsets[3]++);
+        break;
+      case 1:
+        col_oid_to_id[column.GetOid()] = col_id_t(offsets[4]++);
+        break;
+      default:
+        throw std::runtime_error("unexpected switch case value");
+    }
+  }
+
+  auto layout = storage::BlockLayout(attr_sizes);
+  table_ = {new DataTable(block_store_, layout, layout_version_t(0)), layout, col_oid_to_id};
+}
 
 std::vector<col_id_t> SqlTable::ColIdsForOids(const std::vector<catalog::col_oid_t> &col_oids) const {
   TERRIER_ASSERT(!col_oids.empty(), "Should be used to access at least one column.");
